Preferences operations succeed when thread is interrupted.
FutureBarrier waits for Futures, which throw InterruptedException if thread is interrupted.
If such a case, the exception is passed to exception handler, while methods like "completeBlockingWihResult" return defaultValue instead of actual value. This causes inconvenience on library user's side: if thread is interrupted, the returned value is incorrect, and it's hard to detect that the exact call to get value returned an incorrect value.

In similar cases, the original SharedPreferencesImpl loops waiting for condition and swallowing any InterruptedExceptions, so it returns the actual value even if the thread is interrupted.

Getting the future uninterruptibly makes binaryprefs behavior closer to the behavior of original SharedPreferences.
However, while SharedPreferencesImpl simply swallows the InterruptedException, the "uninterruptible get" sets the interrupted status back to the thread.